update to latest node snapshot

+++ consensus-client-it/src/test/scala/units/ManyTransfersTestSuite.scala
@@ -183,7 +183,7 @@ class ManyTransfersTestSuite extends BaseDockerTestSuite {
     waves1.api.broadcastAndWait(
       TxHelpers.dataEntry(
         chainContractAccount,
-        IntegerDataEntry("strictC2ETransfersActivationEpoch", activationEpoch)
+        IntegerDataEntry("strictC2ETransfersActivationEpoch", activationEpoch.toInt)
       )
     )
