+++ consensus-client-it/src/test/scala/units/block/validation/BaseBlockValidationSuite.scala
@@ -123,7 +123,7 @@ trait BaseBlockValidationSuite extends BaseDockerTestSuite {
     waves1.api.broadcastAndWait(
       TxHelpers.dataEntry(
         chainContractAccount,
-        IntegerDataEntry("strictC2ETransfersActivationEpoch", activationEpoch)
+        IntegerDataEntry("strictC2ETransfersActivationEpoch", activationEpoch.toInt)
       )
     )
